requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=2.1.0=pypi_0
aiobotocore=2.9.0=pypi_0
aiohttp=3.9.1=pypi_0
aioitertools=0.11.0=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.2.0=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=2.120=mkl
blas-devel=3.9.0=20_osx64_mkl
blessed=1.20.0=pypi_0
boto3=1.33.13=pypi_0
botocore=1.33.13=pypi_0
brotli-python=1.1.0=py311hdf8f085_1
bzip2=1.0.8=h10d778d_5
ca-certificates=2023.11.17=h8857fd0_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
croniter=1.4.1=pypi_0
cycler=0.12.1=pypi_0
dateutils=0.6.12=pypi_0
debugpy=1.8.0=py311hdf8f085_1
decorator=5.1.1=pyhd8ed1ab_0
deepdiff=6.7.1=pypi_0
docker=6.1.3=pypi_0
docstring-parser=0.15=pypi_0
editor=1.6.5=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.109.0=pypi_0
ffmpeg=4.3=h0a44026_0
filelock=3.13.1=pyhd8ed1ab_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=h60636b9_2
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
gast=0.5.4=pypi_0
gettext=0.21.1=h8a4c099_0
gmp=6.3.0=h93d8f39_0
gmpy2=2.1.2=py311hc5b4402_1
gnutls=3.6.13=h756fd2b_1
google-auth=2.26.2=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
hydra-core=1.3.2=pypi_0
icu=73.2=hf5e326d_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.1=pyha770c72_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.0.1=hd8ed1ab_0
inquirer=3.2.1=pypi_0
ipykernel=6.28.0=pyh3cd1d5f_0
ipython=8.20.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
jmespath=1.0.1=pypi_0
jsonargparse=4.27.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py311h6eed73b_0
keras=2.15.0=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=hb7f2c08_1003
lcms2=2.16=ha2f27b4_0
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=20_osx64_mkl
libcblas=3.9.0=20_osx64_mkl
libclang=16.0.6=pypi_0
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.19=ha4e1b8e_0
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_1
libgfortran5=13.2.0=h2873a65_1
libhwloc=2.9.3=default_h24e0189_1009
libiconv=1.17=hd75f5a5_2
libjpeg-turbo=3.0.0=h0dc2134_1
liblapack=3.9.0=20_osx64_mkl
liblapacke=3.9.0=20_osx64_mkl
libpng=1.6.39=ha978bb4_0
libsodium=1.0.18=hbcb3906_1
libsqlite=3.44.2=h92b6c6a_0
libtiff=4.6.0=h684deea_2
libwebp-base=1.3.2=h0dc2134_0
libxcb=1.15=hb7f2c08_0
libxml2=2.12.4=hc0ae0f7_1
libzlib=1.2.13=h8a1eda9_5
lightning=2.1.3=pypi_0
lightning-api-access=0.0.5=pypi_0
lightning-cloud=0.5.57=pypi_0
lightning-fabric=2.1.3=pypi_0
lightning-utilities=0.10.0=pypi_0
llvm-openmp=17.0.6=hb6ac08f_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py311h2725bcf_1
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.2.0=h54c2260_50500
mkl-devel=2023.2.0=h694c41f_50500
mkl-include=2023.2.0=h6bab518_50500
ml-dtypes=0.2.0=pypi_0
mpc=1.3.1=h81bd1dd_0
mpfr=4.2.1=h0c69b56_0
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.4=pypi_0
ncurses=6.4=h93d8f39_2
nest-asyncio=1.5.9=pyhd8ed1ab_0
nettle=3.6=hedd7734_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.3=py311hc43a94b_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
openh264=2.1.1=hfd3ada9_0
openjpeg=2.5.0=ha4da562_3
openssl=3.2.0=hd75f5a5_1
opt-einsum=3.3.0=pypi_0
ordered-set=4.1.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.2.0=py311hea5c87a_0
pip=23.3.2=pyhd8ed1ab_0
platformdirs=4.1.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.23.4=pypi_0
psutil=5.9.7=py311he705e18_0
pthread-stubs=0.4=hc929b4f_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.7=h9f0c242_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-multipart=0.0.6=pypi_0
python_abi=3.11=4_cp311
pytorch=2.1.2=py3.11_0
pytorch-lightning=2.1.3=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=py311h2725bcf_1
pyzmq=25.1.2=py311h889d6d6_0
readchar=4.0.5=pypi_0
readline=8.2=h9e318b2_1
redis=5.0.1=pypi_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rich=13.7.0=pypi_0
rsa=4.9=pypi_0
runs=1.2.0=pypi_0
s3fs=2023.12.2=pypi_0
s3transfer=0.8.2=pypi_0
scipy=1.11.4=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.35.1=pypi_0
sympy=1.12=pypyh9d50eac_103
tbb=2021.11.0=he51d815_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.35.0=pypi_0
termcolor=2.4.0=pypi_0
tk=8.6.13=h1abcd95_1
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.1.2=py311_cpu
torchdata=0.7.1=pypi_0
torchmetrics=1.3.0=pypi_0
torchtext=0.16.2=pypi_0
torchvision=0.16.2=py311_cpu
tornado=6.3.3=py311h2725bcf_1
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
types-python-dateutil=2.8.19.20240106=pypi_0
typeshed-client=2.4.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
urllib3=2.0.7=pypi_0
uvicorn=0.26.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websocket-client=1.7.0=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xmod=1.8.1=pypi_0
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
yaml=0.2.5=h0d85af4_2
yarl=1.9.4=pypi_0
zeromq=4.3.5=h93d8f39_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8a1eda9_5
zstd=1.5.5=h829000d_0
